# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/roles": {
      "get": {
        "operationId": "UserController_getUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lấy vai trò theo người dùng",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UserController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Gán vai trò cho người dùng",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Gỡ vai trò khỏi người dùng",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "REST Service",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}